requirements.txt
# Core ML
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# Diffusers / HuggingFace
diffusers==0.29.2
transformers==4.41.2
accelerate==0.30.0
safetensors==0.4.3
huggingface-hub==0.23.2
xformers==0.0.27

# IP-Adapter (install from source)
ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git

# InsightFace
insightface==0.7.3
onnxruntime-gpu==1.18.0

# GFPGAN
gfpgan==1.3.8
basicsr==1.4.2
facexlib==0.3.0
realesrgan==0.3.0

# Computer Vision
opencv-python==4.10.0.82
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.1

# UI
streamlit==1.35.0

# Utilities
requests==2.32.3
tqdm==4.66.4
